requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.7
Pillow==8.3.2
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-cleanup==6.0.0
django-resized==1.0.2
mysqlclient==2.1.1